@@ -1,6 +1,6 @@
 module github.com/charmbracelet/x/cellbuf
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/charmbracelet/colorprofile v0.3.2
