@@ -1,8 +1,8 @@
 module github.com/charmbracelet/x/conpty
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/charmbracelet/x/errors v0.0.0-20240508181413-e8d8b6e2de86
-	golang.org/x/sys v0.35.0
+	golang.org/x/sys v0.36.0
 )
