@@ -1,3 +1,3 @@
 module github.com/charmbracelet/x/editor
 
-go 1.23.0
+go 1.24.0
