@@ -1,5 +1,8 @@
 module github.com/charmbracelet/x/exp/charmtone
 
-go 1.23.0
+go 1.24.0
 
-require github.com/lucasb-eyer/go-colorful v1.2.0
+require (
+	github.com/charmbracelet/x/exp/color v0.0.0-20250915100343-2c2e5896ae6e
+	github.com/lucasb-eyer/go-colorful v1.3.0
+)
