@@ -1,5 +1,5 @@
 module github.com/charmbracelet/x/exp/color
 
-go 1.23.0
+go 1.24.0
 
 require github.com/lucasb-eyer/go-colorful v1.2.0
