@@ -1,5 +1,5 @@
 module github.com/charmbracelet/x/exp/golden
 
-go 1.23.0
+go 1.24.0
 
 require github.com/aymanbagabas/go-udiff v0.3.1
