@@ -7,11 +7,11 @@ import (
 
 // nextTab moves the cursor to the next tab stop n times. This respects the
 // horizontal scrolling region. This performs the same function as [ansi.CHT].
-func (t *Emulator) nextTab(n int) {
-	x, y := t.scr.CursorPosition()
-	scroll := t.scr.ScrollRegion()
+func (e *Emulator) nextTab(n int) {
+	x, y := e.scr.CursorPosition()
+	scroll := e.scr.ScrollRegion()
 	for range n {
-		ts := t.tabstops.Next(x)
+		ts := e.tabstops.Next(x)
 		if ts < x {
 			break
 		}
@@ -19,28 +19,28 @@ func (t *Emulator) nextTab(n int) {
 	}
 
 	if x >= scroll.Max.X {
-		x = min(scroll.Max.X-1, t.Width()-1)
+		x = min(scroll.Max.X-1, e.Width()-1)
 	}
 
 	// NOTE: We use t.scr.setCursor here because we don't want to reset the
 	// phantom state.
-	t.scr.setCursor(x, y, false)
+	e.scr.setCursor(x, y, false)
 }
 
 // prevTab moves the cursor to the previous tab stop n times. This respects the
 // horizontal scrolling region when origin mode is set. If the cursor would
 // move past the leftmost valid column, the cursor remains at the leftmost
 // valid column and the operation completes.
-func (t *Emulator) prevTab(n int) {
-	x, _ := t.scr.CursorPosition()
+func (e *Emulator) prevTab(n int) {
+	x, _ := e.scr.CursorPosition()
 	leftmargin := 0
-	scroll := t.scr.ScrollRegion()
-	if t.isModeSet(ansi.DECOM) {
+	scroll := e.scr.ScrollRegion()
+	if e.isModeSet(ansi.DECOM) {
 		leftmargin = scroll.Min.X
 	}
 
 	for range n {
-		ts := t.tabstops.Prev(x)
+		ts := e.tabstops.Prev(x)
 		if ts > x {
 			break
 		}
@@ -53,29 +53,29 @@ func (t *Emulator) prevTab(n int) {
 
 	// NOTE: We use t.scr.setCursorX here because we don't want to reset the
 	// phantom state.
-	t.scr.setCursorX(x, false)
+	e.scr.setCursorX(x, false)
 }
 
 // moveCursor moves the cursor by the given x and y deltas. If the cursor
 // is at phantom, the state will reset and the cursor is back in the screen.
-func (t *Emulator) moveCursor(dx, dy int) {
-	t.scr.moveCursor(dx, dy)
-	t.atPhantom = false
+func (e *Emulator) moveCursor(dx, dy int) {
+	e.scr.moveCursor(dx, dy)
+	e.atPhantom = false
 }
 
 // setCursor sets the cursor position. This resets the phantom state.
-func (t *Emulator) setCursor(x, y int) {
-	t.scr.setCursor(x, y, false)
-	t.atPhantom = false
+func (e *Emulator) setCursor(x, y int) {
+	e.scr.setCursor(x, y, false)
+	e.atPhantom = false
 }
 
 // setCursorPosition sets the cursor position. This respects [ansi.DECOM],
 // Origin Mode. This performs the same function as [ansi.CUP].
-func (t *Emulator) setCursorPosition(x, y int) {
-	mode, ok := t.modes[ansi.DECOM]
+func (e *Emulator) setCursorPosition(x, y int) {
+	mode, ok := e.modes[ansi.DECOM]
 	margins := ok && mode.IsSet()
-	t.scr.setCursor(x, y, margins)
-	t.atPhantom = false
+	e.scr.setCursor(x, y, margins)
+	e.atPhantom = false
 }
 
 // carriageReturn moves the cursor to the leftmost column. If [ansi.DECOM] is
@@ -83,28 +83,28 @@ func (t *Emulator) setCursorPosition(x, y int) {
 // to the right of the left margin, the cursor is set to the left margin.
 // Otherwise, the cursor is set to the leftmost column of the screen.
 // This performs the same function as [ansi.CR].
-func (t *Emulator) carriageReturn() {
-	mode, ok := t.modes[ansi.DECOM]
+func (e *Emulator) carriageReturn() {
+	mode, ok := e.modes[ansi.DECOM]
 	margins := ok && mode.IsSet()
-	x, y := t.scr.CursorPosition()
+	x, y := e.scr.CursorPosition()
 	if margins {
-		t.scr.setCursor(0, y, true)
-	} else if region := t.scr.ScrollRegion(); uv.Pos(x, y).In(region) {
-		t.scr.setCursor(region.Min.X, y, false)
+		e.scr.setCursor(0, y, true)
+	} else if region := e.scr.ScrollRegion(); uv.Pos(x, y).In(region) {
+		e.scr.setCursor(region.Min.X, y, false)
 	} else {
-		t.scr.setCursor(0, y, false)
+		e.scr.setCursor(0, y, false)
 	}
-	t.atPhantom = false
+	e.atPhantom = false
 }
 
 // repeatPreviousCharacter repeats the previous character n times. This is
 // equivalent to typing the same character n times. This performs the same as
 // [ansi.REP].
-func (t *Emulator) repeatPreviousCharacter(n int) {
-	if t.lastChar == 0 {
+func (e *Emulator) repeatPreviousCharacter(n int) {
+	if e.lastChar == 0 {
 		return
 	}
 	for range n {
-		t.handlePrint(t.lastChar)
+		e.handlePrint(e.lastChar)
 	}
 }
