@@ -6,13 +6,13 @@ import (
 
 // eraseCharacter erases n characters starting from the cursor position. It
 // does not move the cursor. This is equivalent to [ansi.ECH].
-func (t *Emulator) eraseCharacter(n int) {
+func (e *Emulator) eraseCharacter(n int) {
 	if n <= 0 {
 		n = 1
 	}
-	x, y := t.scr.CursorPosition()
+	x, y := e.scr.CursorPosition()
 	rect := uv.Rect(x, y, n, 1)
-	t.scr.FillArea(t.scr.blankCell(), rect)
-	t.atPhantom = false
+	e.scr.FillArea(e.scr.blankCell(), rect)
+	e.atPhantom = false
 	// ECH does not move the cursor.
 }
