@@ -211,50 +211,50 @@ func (h *handlers) handleEsc(cmd int) bool {
 }
 
 // registerDefaultHandlers registers the default escape sequence handlers.
-func (t *Emulator) registerDefaultHandlers() {
-	t.registerDefaultCcHandlers()
-	t.registerDefaultCsiHandlers()
-	t.registerDefaultEscHandlers()
-	t.registerDefaultOscHandlers()
+func (e *Emulator) registerDefaultHandlers() {
+	e.registerDefaultCcHandlers()
+	e.registerDefaultCsiHandlers()
+	e.registerDefaultEscHandlers()
+	e.registerDefaultOscHandlers()
 }
 
 // registerDefaultCcHandlers registers the default control character handlers.
-func (t *Emulator) registerDefaultCcHandlers() {
+func (e *Emulator) registerDefaultCcHandlers() {
 	for i := byte(ansi.NUL); i <= ansi.US; i++ {
 		switch i {
 		case ansi.NUL: // Null [ansi.NUL]
 			// Ignored
-			t.registerCcHandler(i, func() bool {
+			e.registerCcHandler(i, func() bool {
 				return true
 			})
 		case ansi.BEL: // Bell [ansi.BEL]
-			t.registerCcHandler(i, func() bool {
-				if t.cb.Bell != nil {
-					t.cb.Bell()
+			e.registerCcHandler(i, func() bool {
+				if e.cb.Bell != nil {
+					e.cb.Bell()
 				}
 				return true
 			})
 		case ansi.BS: // Backspace [ansi.BS]
-			t.registerCcHandler(i, func() bool {
-				t.backspace()
+			e.registerCcHandler(i, func() bool {
+				e.backspace()
 				return true
 			})
 		case ansi.HT: // Horizontal Tab [ansi.HT]
-			t.registerCcHandler(i, func() bool {
-				t.nextTab(1)
+			e.registerCcHandler(i, func() bool {
+				e.nextTab(1)
 				return true
 			})
 		case ansi.LF, ansi.VT, ansi.FF:
 			// Line Feed [ansi.LF]
 			// Vertical Tab [ansi.VT]
 			// Form Feed [ansi.FF]
-			t.registerCcHandler(i, func() bool {
-				t.linefeed()
+			e.registerCcHandler(i, func() bool {
+				e.linefeed()
 				return true
 			})
 		case ansi.CR: // Carriage Return [ansi.CR]
-			t.registerCcHandler(i, func() bool {
-				t.carriageReturn()
+			e.registerCcHandler(i, func() bool {
+				e.carriageReturn()
 				return true
 			})
 		}
@@ -263,38 +263,38 @@ func (t *Emulator) registerDefaultCcHandlers() {
 	for i := byte(ansi.PAD); i <= byte(ansi.APC); i++ {
 		switch i {
 		case ansi.HTS: // Horizontal Tab Set [ansi.HTS]
-			t.registerCcHandler(i, func() bool {
-				t.horizontalTabSet()
+			e.registerCcHandler(i, func() bool {
+				e.horizontalTabSet()
 				return true
 			})
 		case ansi.RI: // Reverse Index [ansi.RI]
-			t.registerCcHandler(i, func() bool {
-				t.reverseIndex()
+			e.registerCcHandler(i, func() bool {
+				e.reverseIndex()
 				return true
 			})
 		case ansi.SO: // Shift Out [ansi.SO]
-			t.registerCcHandler(i, func() bool {
-				t.gl = 1
+			e.registerCcHandler(i, func() bool {
+				e.gl = 1
 				return true
 			})
 		case ansi.SI: // Shift In [ansi.SI]
-			t.registerCcHandler(i, func() bool {
-				t.gl = 0
+			e.registerCcHandler(i, func() bool {
+				e.gl = 0
 				return true
 			})
 		case ansi.IND: // Index [ansi.IND]
-			t.registerCcHandler(i, func() bool {
-				t.index()
+			e.registerCcHandler(i, func() bool {
+				e.index()
 				return true
 			})
 		case ansi.SS2: // Single Shift 2 [ansi.SS2]
-			t.registerCcHandler(i, func() bool {
-				t.gsingle = 2
+			e.registerCcHandler(i, func() bool {
+				e.gsingle = 2
 				return true
 			})
 		case ansi.SS3: // Single Shift 3 [ansi.SS3]
-			t.registerCcHandler(i, func() bool {
-				t.gsingle = 3
+			e.registerCcHandler(i, func() bool {
+				e.gsingle = 3
 				return true
 			})
 		}
@@ -302,28 +302,28 @@ func (t *Emulator) registerDefaultCcHandlers() {
 }
 
 // registerDefaultOscHandlers registers the default OSC escape sequence handlers.
-func (t *Emulator) registerDefaultOscHandlers() {
+func (e *Emulator) registerDefaultOscHandlers() {
 	for _, cmd := range []int{
 		0, // Set window title and icon name
 		1, // Set icon name
 		2, // Set window title
 	} {
-		t.RegisterOscHandler(cmd, func(data []byte) bool {
-			t.handleTitle(cmd, data)
+		e.RegisterOscHandler(cmd, func(data []byte) bool {
+			e.handleTitle(cmd, data)
 			return true
 		})
 	}
 
-	t.RegisterOscHandler(7, func(data []byte) bool {
+	e.RegisterOscHandler(7, func(data []byte) bool {
 		// Report the shell current working directory
 		// [ansi.NotifyWorkingDirectory].
-		t.handleWorkingDirectory(7, data)
+		e.handleWorkingDirectory(7, data)
 		return true
 	})
 
-	t.RegisterOscHandler(8, func(data []byte) bool {
+	e.RegisterOscHandler(8, func(data []byte) bool {
 		// Set/Query Hyperlink [ansi.SetHyperlink]
-		t.handleHyperlink(8, data)
+		e.handleHyperlink(8, data)
 		return true
 	})
 
@@ -335,36 +335,36 @@ func (t *Emulator) registerDefaultOscHandlers() {
 		111, // Reset background color
 		112, // Reset cursor color
 	} {
-		t.RegisterOscHandler(cmd, func(data []byte) bool {
-			t.handleDefaultColor(cmd, data)
+		e.RegisterOscHandler(cmd, func(data []byte) bool {
+			e.handleDefaultColor(cmd, data)
 			return true
 		})
 	}
 }
 
 // registerDefaultEscHandlers registers the default ESC escape sequence handlers.
-func (t *Emulator) registerDefaultEscHandlers() {
-	t.RegisterEscHandler('=', func() bool {
+func (e *Emulator) registerDefaultEscHandlers() {
+	e.RegisterEscHandler('=', func() bool {
 		// Keypad Application Mode [ansi.DECKPAM]
-		t.setMode(ansi.NumericKeypadMode, ansi.ModeSet)
+		e.setMode(ansi.NumericKeypadMode, ansi.ModeSet)
 		return true
 	})
 
-	t.RegisterEscHandler('>', func() bool {
+	e.RegisterEscHandler('>', func() bool {
 		// Keypad Numeric Mode [ansi.DECKPNM]
-		t.setMode(ansi.NumericKeypadMode, ansi.ModeReset)
+		e.setMode(ansi.NumericKeypadMode, ansi.ModeReset)
 		return true
 	})
 
-	t.RegisterEscHandler('7', func() bool {
+	e.RegisterEscHandler('7', func() bool {
 		// Save Cursor [ansi.DECSC]
-		t.scr.SaveCursor()
+		e.scr.SaveCursor()
 		return true
 	})
 
-	t.RegisterEscHandler('8', func() bool {
+	e.RegisterEscHandler('8', func() bool {
 		// Restore Cursor [ansi.DECRC]
-		t.scr.RestoreCursor()
+		e.scr.RestoreCursor()
 		return true
 	})
 
@@ -382,17 +382,17 @@ func (t *Emulator) registerDefaultEscHandlers() {
 		ansi.Command(0, '*', '0'), // Special G2
 		ansi.Command(0, '+', '0'), // Special G3
 	} {
-		t.RegisterEscHandler(cmd, func() bool {
+		e.RegisterEscHandler(cmd, func() bool {
 			// Select Character Set [ansi.SCS]
 			c := ansi.Cmd(cmd)
 			set := c.Intermediate() - '('
 			switch c.Final() {
 			case 'A': // UK Character Set
-				t.charsets[set] = UK
+				e.charsets[set] = UK
 			case 'B': // USASCII Character Set
-				t.charsets[set] = nil // USASCII is the default
+				e.charsets[set] = nil // USASCII is the default
 			case '0': // Special Drawing Character Set
-				t.charsets[set] = SpecialDrawing
+				e.charsets[set] = SpecialDrawing
 			default:
 				return false
 			}
@@ -400,125 +400,125 @@ func (t *Emulator) registerDefaultEscHandlers() {
 		})
 	}
 
-	t.RegisterEscHandler('D', func() bool {
+	e.RegisterEscHandler('D', func() bool {
 		// Index [ansi.IND]
-		t.index()
+		e.index()
 		return true
 	})
 
-	t.RegisterEscHandler('H', func() bool {
+	e.RegisterEscHandler('H', func() bool {
 		// Horizontal Tab Set [ansi.HTS]
-		t.horizontalTabSet()
+		e.horizontalTabSet()
 		return true
 	})
 
-	t.RegisterEscHandler('M', func() bool {
+	e.RegisterEscHandler('M', func() bool {
 		// Reverse Index [ansi.RI]
-		t.reverseIndex()
+		e.reverseIndex()
 		return true
 	})
 
-	t.RegisterEscHandler('c', func() bool {
+	e.RegisterEscHandler('c', func() bool {
 		// Reset Initial State [ansi.RIS]
-		t.fullReset()
+		e.fullReset()
 		return true
 	})
 
-	t.RegisterEscHandler('n', func() bool {
+	e.RegisterEscHandler('n', func() bool {
 		// Locking Shift G2 [ansi.LS2]
-		t.gl = 2
+		e.gl = 2
 		return true
 	})
 
-	t.RegisterEscHandler('o', func() bool {
+	e.RegisterEscHandler('o', func() bool {
 		// Locking Shift G3 [ansi.LS3]
-		t.gl = 3
+		e.gl = 3
 		return true
 	})
 
-	t.RegisterEscHandler('|', func() bool {
+	e.RegisterEscHandler('|', func() bool {
 		// Locking Shift 3 Right [ansi.LS3R]
-		t.gr = 3
+		e.gr = 3
 		return true
 	})
 
-	t.RegisterEscHandler('}', func() bool {
+	e.RegisterEscHandler('}', func() bool {
 		// Locking Shift 2 Right [ansi.LS2R]
-		t.gr = 2
+		e.gr = 2
 		return true
 	})
 
-	t.RegisterEscHandler('~', func() bool {
+	e.RegisterEscHandler('~', func() bool {
 		// Locking Shift 1 Right [ansi.LS1R]
-		t.gr = 1
+		e.gr = 1
 		return true
 	})
 }
 
 // registerDefaultCsiHandlers registers the default CSI escape sequence handlers.
-func (t *Emulator) registerDefaultCsiHandlers() {
-	t.RegisterCsiHandler('@', func(params ansi.Params) bool {
+func (e *Emulator) registerDefaultCsiHandlers() {
+	e.RegisterCsiHandler('@', func(params ansi.Params) bool {
 		// Insert Character [ansi.ICH]
 		n, _, _ := params.Param(0, 1)
-		t.scr.InsertCell(n)
+		e.scr.InsertCell(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('A', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('A', func(params ansi.Params) bool {
 		// Cursor Up [ansi.CUU]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(0, -n)
+		e.moveCursor(0, -n)
 		return true
 	})
 
-	t.RegisterCsiHandler('B', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('B', func(params ansi.Params) bool {
 		// Cursor Down [ansi.CUD]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(0, n)
+		e.moveCursor(0, n)
 		return true
 	})
 
-	t.RegisterCsiHandler('C', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('C', func(params ansi.Params) bool {
 		// Cursor Forward [ansi.CUF]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(n, 0)
+		e.moveCursor(n, 0)
 		return true
 	})
 
-	t.RegisterCsiHandler('D', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('D', func(params ansi.Params) bool {
 		// Cursor Backward [ansi.CUB]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(-n, 0)
+		e.moveCursor(-n, 0)
 		return true
 	})
 
-	t.RegisterCsiHandler('E', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('E', func(params ansi.Params) bool {
 		// Cursor Next Line [ansi.CNL]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(0, n)
-		t.carriageReturn()
+		e.moveCursor(0, n)
+		e.carriageReturn()
 		return true
 	})
 
-	t.RegisterCsiHandler('F', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('F', func(params ansi.Params) bool {
 		// Cursor Previous Line [ansi.CPL]
 		n, _, _ := params.Param(0, 1)
-		t.moveCursor(0, -n)
-		t.carriageReturn()
+		e.moveCursor(0, -n)
+		e.carriageReturn()
 		return true
 	})
 
-	t.RegisterCsiHandler('G', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('G', func(params ansi.Params) bool {
 		// Cursor Horizontal Absolute [ansi.CHA]
 		n, _, _ := params.Param(0, 1)
-		_, y := t.scr.CursorPosition()
-		t.setCursor(n-1, y)
+		_, y := e.scr.CursorPosition()
+		e.setCursor(n-1, y)
 		return true
 	})
 
-	t.RegisterCsiHandler('H', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('H', func(params ansi.Params) bool {
 		// Cursor Position [ansi.CUP]
-		width, height := t.Width(), t.Height()
+		width, height := e.Width(), e.Height()
 		row, _, _ := params.Param(0, 1)
 		col, _, _ := params.Param(1, 1)
 		if row < 1 {
@@ -529,158 +529,158 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		}
 		y := min(height-1, row-1)
 		x := min(width-1, col-1)
-		t.setCursorPosition(x, y)
+		e.setCursorPosition(x, y)
 		return true
 	})
 
-	t.RegisterCsiHandler('I', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('I', func(params ansi.Params) bool {
 		// Cursor Horizontal Tabulation [ansi.CHT]
 		n, _, _ := params.Param(0, 1)
-		t.nextTab(n)
+		e.nextTab(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('J', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('J', func(params ansi.Params) bool {
 		// Erase in Display [ansi.ED]
 		n, _, _ := params.Param(0, 0)
-		width, height := t.Width(), t.Height()
-		x, y := t.scr.CursorPosition()
+		width, height := e.Width(), e.Height()
+		x, y := e.scr.CursorPosition()
 		switch n {
 		case 0: // Erase screen below (from after cursor position)
 			rect1 := uv.Rect(x, y, width, 1)            // cursor to end of line
 			rect2 := uv.Rect(0, y+1, width, height-y-1) // next line onwards
-			t.scr.FillArea(t.scr.blankCell(), rect1)
-			t.scr.FillArea(t.scr.blankCell(), rect2)
+			e.scr.FillArea(e.scr.blankCell(), rect1)
+			e.scr.FillArea(e.scr.blankCell(), rect2)
 		case 1: // Erase screen above (including cursor)
 			rect := uv.Rect(0, 0, width, y+1)
-			t.scr.FillArea(t.scr.blankCell(), rect)
+			e.scr.FillArea(e.scr.blankCell(), rect)
 		case 2: // erase screen
 			fallthrough
 		case 3: // erase display
 			//nolint:godox
 			// TODO: Scrollback buffer support?
-			t.scr.Clear()
+			e.scr.Clear()
 		default:
 			return false
 		}
 		return true
 	})
 
-	t.RegisterCsiHandler('K', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('K', func(params ansi.Params) bool {
 		// Erase in Line [ansi.EL]
 		n, _, _ := params.Param(0, 0)
 		// NOTE: Erase Line (EL) erases all character attributes but not cell
 		// bg color.
-		x, y := t.scr.CursorPosition()
-		w := t.scr.Width()
+		x, y := e.scr.CursorPosition()
+		w := e.scr.Width()
 
 		switch n {
 		case 0: // Erase from cursor to end of line
-			t.eraseCharacter(w - x)
+			e.eraseCharacter(w - x)
 		case 1: // Erase from start of line to cursor
 			rect := uv.Rect(0, y, x+1, 1)
-			t.scr.FillArea(t.scr.blankCell(), rect)
+			e.scr.FillArea(e.scr.blankCell(), rect)
 		case 2: // Erase entire line
 			rect := uv.Rect(0, y, w, 1)
-			t.scr.FillArea(t.scr.blankCell(), rect)
+			e.scr.FillArea(e.scr.blankCell(), rect)
 		default:
 			return false
 		}
 		return true
 	})
 
-	t.RegisterCsiHandler('L', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('L', func(params ansi.Params) bool {
 		// Insert Line [ansi.IL]
 		n, _, _ := params.Param(0, 1)
-		if t.scr.InsertLine(n) {
+		if e.scr.InsertLine(n) {
 			// Move the cursor to the left margin.
-			t.scr.setCursorX(0, true)
+			e.scr.setCursorX(0, true)
 		}
 		return true
 	})
 
-	t.RegisterCsiHandler('M', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('M', func(params ansi.Params) bool {
 		// Delete Line [ansi.DL]
 		n, _, _ := params.Param(0, 1)
-		if t.scr.DeleteLine(n) {
+		if e.scr.DeleteLine(n) {
 			// If the line was deleted successfully, move the cursor to the
 			// left.
 			// Move the cursor to the left margin.
-			t.scr.setCursorX(0, true)
+			e.scr.setCursorX(0, true)
 		}
 		return true
 	})
 
-	t.RegisterCsiHandler('P', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('P', func(params ansi.Params) bool {
 		// Delete Character [ansi.DCH]
 		n, _, _ := params.Param(0, 1)
-		t.scr.DeleteCell(n)
+		e.scr.DeleteCell(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('S', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('S', func(params ansi.Params) bool {
 		// Scroll Up [ansi.SU]
 		n, _, _ := params.Param(0, 1)
-		t.scr.ScrollUp(n)
+		e.scr.ScrollUp(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('T', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('T', func(params ansi.Params) bool {
 		// Scroll Down [ansi.SD]
 		n, _, _ := params.Param(0, 1)
-		t.scr.ScrollDown(n)
+		e.scr.ScrollDown(n)
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('?', 0, 'W'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('?', 0, 'W'), func(params ansi.Params) bool {
 		// Set Tab at Every 8 Columns [ansi.DECST8C]
 		if len(params) == 1 && params[0] == 5 {
-			t.resetTabStops()
+			e.resetTabStops()
 			return true
 		}
 		return false
 	})
 
-	t.RegisterCsiHandler('X', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('X', func(params ansi.Params) bool {
 		// Erase Character [ansi.ECH]
 		n, _, _ := params.Param(0, 1)
-		t.eraseCharacter(n)
+		e.eraseCharacter(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('Z', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('Z', func(params ansi.Params) bool {
 		// Cursor Backward Tabulation [ansi.CBT]
 		n, _, _ := params.Param(0, 1)
-		t.prevTab(n)
+		e.prevTab(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('`', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('`', func(params ansi.Params) bool {
 		// Horizontal Position Absolute [ansi.HPA]
 		n, _, _ := params.Param(0, 1)
-		width := t.Width()
-		_, y := t.scr.CursorPosition()
-		t.setCursorPosition(min(width-1, n-1), y)
+		width := e.Width()
+		_, y := e.scr.CursorPosition()
+		e.setCursorPosition(min(width-1, n-1), y)
 		return true
 	})
 
-	t.RegisterCsiHandler('a', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('a', func(params ansi.Params) bool {
 		// Horizontal Position Relative [ansi.HPR]
 		n, _, _ := params.Param(0, 1)
-		width := t.Width()
-		x, y := t.scr.CursorPosition()
-		t.setCursorPosition(min(width-1, x+n), y)
+		width := e.Width()
+		x, y := e.scr.CursorPosition()
+		e.setCursorPosition(min(width-1, x+n), y)
 		return true
 	})
 
-	t.RegisterCsiHandler('b', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('b', func(params ansi.Params) bool {
 		// Repeat Previous Character [ansi.REP]
 		n, _, _ := params.Param(0, 1)
-		t.repeatPreviousCharacter(n)
+		e.repeatPreviousCharacter(n)
 		return true
 	})
 
-	t.RegisterCsiHandler('c', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('c', func(params ansi.Params) bool {
 		// Primary Device Attributes [ansi.DA1]
 		n, _, _ := params.Param(0, 0)
 		if n != 0 {
@@ -688,7 +688,7 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		}
 
 		// Do we fully support VT220?
-		_, _ = io.WriteString(t.pw, ansi.PrimaryDeviceAttributes(
+		_, _ = io.WriteString(e.pw, ansi.PrimaryDeviceAttributes(
 			62, // VT220
 			1,  // 132 columns
 			6,  // Selective Erase
@@ -697,7 +697,7 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('>', 0, 'c'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('>', 0, 'c'), func(params ansi.Params) bool {
 		// Secondary Device Attributes [ansi.DA2]
 		n, _, _ := params.Param(0, 0)
 		if n != 0 {
@@ -705,7 +705,7 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		}
 
 		// Do we fully support VT220?
-		_, _ = io.WriteString(t.pw, ansi.SecondaryDeviceAttributes(
+		_, _ = io.WriteString(e.pw, ansi.SecondaryDeviceAttributes(
 			1,  // VT220
 			10, // Version 1.0
 			0,  // ROM Cartridge is always zero
@@ -713,44 +713,44 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		return true
 	})
 
-	t.RegisterCsiHandler('d', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('d', func(params ansi.Params) bool {
 		// Vertical Position Absolute [ansi.VPA]
 		n, _, _ := params.Param(0, 1)
-		height := t.Height()
-		x, _ := t.scr.CursorPosition()
-		t.setCursorPosition(x, min(height-1, n-1))
+		height := e.Height()
+		x, _ := e.scr.CursorPosition()
+		e.setCursorPosition(x, min(height-1, n-1))
 		return true
 	})
 
-	t.RegisterCsiHandler('e', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('e', func(params ansi.Params) bool {
 		// Vertical Position Relative [ansi.VPR]
 		n, _, _ := params.Param(0, 1)
-		height := t.Height()
-		x, y := t.scr.CursorPosition()
-		t.setCursorPosition(x, min(height-1, y+n))
+		height := e.Height()
+		x, y := e.scr.CursorPosition()
+		e.setCursorPosition(x, min(height-1, y+n))
 		return true
 	})
 
-	t.RegisterCsiHandler('f', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('f', func(params ansi.Params) bool {
 		// Horizontal and Vertical Position [ansi.HVP]
-		width, height := t.Width(), t.Height()
+		width, height := e.Width(), e.Height()
 		row, _, _ := params.Param(0, 1)
 		col, _, _ := params.Param(1, 1)
 		y := min(height-1, row-1)
 		x := min(width-1, col-1)
-		t.setCursor(x, y)
+		e.setCursor(x, y)
 		return true
 	})
 
-	t.RegisterCsiHandler('g', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('g', func(params ansi.Params) bool {
 		// Tab Clear [ansi.TBC]
 		value, _, _ := params.Param(0, 0)
 		switch value {
 		case 0:
-			x, _ := t.scr.CursorPosition()
-			t.tabstops.Reset(x)
+			x, _ := e.scr.CursorPosition()
+			e.tabstops.Reset(x)
 		case 3:
-			t.tabstops.Clear()
+			e.tabstops.Clear()
 		default:
 			return false
 		}
@@ -758,37 +758,37 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		return true
 	})
 
-	t.RegisterCsiHandler('h', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('h', func(params ansi.Params) bool {
 		// Set Mode [ansi.SM] - ANSI
-		t.handleMode(params, true, true)
+		e.handleMode(params, true, true)
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('?', 0, 'h'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('?', 0, 'h'), func(params ansi.Params) bool {
 		// Set Mode [ansi.SM] - DEC
-		t.handleMode(params, true, false)
+		e.handleMode(params, true, false)
 		return true
 	})
 
-	t.RegisterCsiHandler('l', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('l', func(params ansi.Params) bool {
 		// Reset Mode [ansi.RM] - ANSI
-		t.handleMode(params, false, true)
+		e.handleMode(params, false, true)
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('?', 0, 'l'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('?', 0, 'l'), func(params ansi.Params) bool {
 		// Reset Mode [ansi.RM] - DEC
-		t.handleMode(params, false, false)
+		e.handleMode(params, false, false)
 		return true
 	})
 
-	t.RegisterCsiHandler('m', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('m', func(params ansi.Params) bool {
 		// Select Graphic Rendition [ansi.SGR]
-		t.handleSgr(params)
+		e.handleSgr(params)
 		return true
 	})
 
-	t.RegisterCsiHandler('n', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('n', func(params ansi.Params) bool {
 		// Device Status Report [ansi.DSR]
 		n, _, ok := params.Param(0, 1)
 		if !ok || n == 0 {
@@ -799,10 +799,10 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		case 5: // Operating Status
 			// We're always ready ;)
 			// See: https://vt100.net/docs/vt510-rm/DSR-OS.html
-			_, _ = io.WriteString(t.pw, ansi.DeviceStatusReport(ansi.DECStatusReport(0)))
+			_, _ = io.WriteString(e.pw, ansi.DeviceStatusReport(ansi.DECStatusReport(0)))
 		case 6: // Cursor Position Report [ansi.CPR]
-			x, y := t.scr.CursorPosition()
-			_, _ = io.WriteString(t.pw, ansi.CursorPositionReport(x+1, y+1))
+			x, y := e.scr.CursorPosition()
+			_, _ = io.WriteString(e.pw, ansi.CursorPositionReport(x+1, y+1))
 		default:
 			return false
 		}
@@ -810,7 +810,7 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('?', 0, 'n'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('?', 0, 'n'), func(params ansi.Params) bool {
 		n, _, ok := params.Param(0, 1)
 		if !ok || n == 0 {
 			return false
@@ -818,8 +818,8 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 
 		switch n {
 		case 6: // Extended Cursor Position Report [ansi.DECXCPR]
-			x, y := t.scr.CursorPosition()
-			_, _ = io.WriteString(t.pw, ansi.ExtendedCursorPositionReport(x+1, y+1, 0)) // We don't support page numbers //nolint:errcheck
+			x, y := e.scr.CursorPosition()
+			_, _ = io.WriteString(e.pw, ansi.ExtendedCursorPositionReport(x+1, y+1, 0)) // We don't support page numbers //nolint:errcheck
 		default:
 			return false
 		}
@@ -827,19 +827,19 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command(0, '$', 'p'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command(0, '$', 'p'), func(params ansi.Params) bool {
 		// Request Mode [ansi.DECRQM] - ANSI
-		t.handleRequestMode(params, true)
+		e.handleRequestMode(params, true)
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command('?', '$', 'p'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command('?', '$', 'p'), func(params ansi.Params) bool {
 		// Request Mode [ansi.DECRQM] - DEC
-		t.handleRequestMode(params, false)
+		e.handleRequestMode(params, false)
 		return true
 	})
 
-	t.RegisterCsiHandler(ansi.Command(0, ' ', 'q'), func(params ansi.Params) bool {
+	e.RegisterCsiHandler(ansi.Command(0, ' ', 'q'), func(params ansi.Params) bool {
 		// Set Cursor Style [ansi.DECSCUSR]
 		n := 1
 		if param, _, ok := params.Param(0, 0); ok && param > n {
@@ -850,19 +850,19 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 		if !blink {
 			style--
 		}
-		t.scr.setCursorStyle(CursorStyle(style), blink)
+		e.scr.setCursorStyle(CursorStyle(style), blink)
 		return true
 	})
 
-	t.RegisterCsiHandler('r', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('r', func(params ansi.Params) bool {
 		// Set Top and Bottom Margins [ansi.DECSTBM]
 		top, _, _ := params.Param(0, 1)
 		if top < 1 {
 			top = 1
 		}
 
-		height := t.Height()
-		bottom, _ := t.parser.Param(1, height)
+		height := e.Height()
+		bottom, _ := e.parser.Param(1, height)
 		if bottom < 1 {
 			bottom = height
 		}
@@ -873,28 +873,28 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 
 		// Rect is [x, y) which means y is exclusive. So the top margin
 		// is the top of the screen minus one.
-		t.scr.setVerticalMargins(top-1, bottom)
+		e.scr.setVerticalMargins(top-1, bottom)
 
 		// Move the cursor to the top-left of the screen or scroll region
 		// depending on [ansi.DECOM].
-		t.setCursorPosition(0, 0)
+		e.setCursorPosition(0, 0)
 		return true
 	})
 
-	t.RegisterCsiHandler('s', func(params ansi.Params) bool {
+	e.RegisterCsiHandler('s', func(params ansi.Params) bool {
 		// Set Left and Right Margins [ansi.DECSLRM]
 		// These conflict with each other. When [ansi.DECSLRM] is set, the we
 		// set the left and right margins. Otherwise, we save the cursor
 		// position.
 
-		if t.isModeSet(ansi.LeftRightMarginMode) {
+		if e.isModeSet(ansi.LeftRightMarginMode) {
 			// Set Left Right Margins [ansi.DECSLRM]
 			left, _, _ := params.Param(0, 1)
 			if left < 1 {
 				left = 1
 			}
 
-			width := t.Width()
+			width := e.Width()
 			right, _, _ := params.Param(1, width)
 			if right < 1 {
 				right = width
@@ -904,14 +904,14 @@ func (t *Emulator) registerDefaultCsiHandlers() {
 				return false
 			}
 
-			t.scr.setHorizontalMargins(left-1, right)
+			e.scr.setHorizontalMargins(left-1, right)
 
 			// Move the cursor to the top-left of the screen or scroll region
 			// depending on [ansi.DECOM].
-			t.setCursorPosition(0, 0)
+			e.setCursorPosition(0, 0)
 		} else {
 			// Save Current Cursor Position [ansi.SCOSC]
-			t.scr.SaveCursor()
+			e.scr.SaveCursor()
 		}
 
 		return true
